restructure with new model design
