+++ sample_model/model.py
@@ -2,7 +2,7 @@
 # LOAD DEPENDENCIES
 # =============================================================================
 
-from modelsandbox import Model
+from modelsandbox_old import Model
 import os, math
 import numpy as np
 MODEL_PATH = os.path.abspath(os.path.dirname(__file__))
